Add a job to import schools data from DfE source
This inserts or updates schools data from the most recent CSV file
available at https://get-information-schools.service.gov.uk/Downloads

A few extra fields were added as part of this:
* `code` added to region and school_types to make lookup more reliable
(i.e. because the label is more likely to change than the code)
* a new ‘detailed school type’ association, which includes values like
“Voluntary aided school” and “Community school” as a breakdown of the
more general “LA maintained school” type.
* additional address fields: `locality` and `address3`.

For now I suggest importing ALL the records, even though some may be
out-of-scope, so that we can filter them later (and display a more
helpful message to the user if they try to list a vacancy for a school
out of scope).