Split vacancy form into 3 parts
Following the design from the prototype, this splits the form into
three page:

1) Job specification
2) Candidate specification
3) Application details

The consequence of this is to make sure that only the appropriate
errors are shown at the right point, and also so that when going back
to step 1 or step 2 from the review page, you then return to the review
rather than stepping through the whole flow all over again.

This also has #29 already merged into it, to avoid having to repeat
some of the work done there which this relies on.