(fix) DfE Sign in redirects will continue to route the user to the sign in page
* This was rendering the user to a white page that said '401 Unauthorised'
* This is because DfE Sign-in will always redirect users to our redirect_uri that’s used for normal authentication requests which in our case is auth/dfe/callback. Without this endpoint DfE Sign-in cannot return us the required information.
* Having more than 1 redirect URL is not configurable in their service and they advise us to handle multiple types of request on the same endpoint.
* This is a problem for us because OmniAuth was identifying this
harmless GET redirect as unauthorised because there was no nonce in the
request, OmniAuth call it `stored_state`. This then is intercepted in
Rack by our Omniauth wrapper that sits around our callback endpoint. To
change the behaviour we have to look inside our omniauth strategy and
add an early guard clause for the situation where the request contains
no query parameters at all. This feels like a reliable enough way to
differentiate the request from the real authentication payload.
* My decision to monkey patch this was balanced against forking the gem
we are using and maintaining it with these 2 small modifications.

# Conflicts:
#	spec/requests/dfe_sign_in_spec.rb