(chore) Existing regions will now use DfE Sign In
As part of the switch to using DfE Sign In for all our users we need to
make sure they see the correct sign in screen.

In a future PR we will remove Azure sign in completely, but for now just
pointing existing options to DfE sign in should suffice.