(refactor) Multiple schools instance variable check count
Checking for the existence of this session key won't tell us if the user
should have access to multiple schools. Instead we should check for the
presence of the key and count the amount of permissions stored. If more
than one school, then we can show the 'Change organisation' link