(feature) Only authorise a selected school if it matches our permissions
We want to have granular control over which organisations have access to
our service. We already track school based permissions in our access
application (TVA). When a user selects a school on DfE Sign In we need
to check that the user has permission for the selected school in TVA.

We still also want to support users who may not have an organisation in
DfE Sign In but do in TVA.