(chore) Explicitly specify the `document_type` on the model. This is required as a result of using elasticsearch gems version 6.x
This is because Elasticsearch has changed the default document type
to "_doc", unless you override it with the document_type setting.
The reason for this change is that the ability to have multiple types
in a single index is deprecated and will be removed in Elasticsearch 7.
Please see the [documentation](https://www.elastic.co/guide/en/elasticsearch/reference/master/removal-of-types.html#_schedule_for_removal_of_mapping_types)