(fix) Only show these warnings in Rails development env
I wanted to test this, perhaps in spec/services/teacher_vacancy_authorisation_spec.rb but it's hard to test initializers with rspec. I can keep plugging away at it if it's important?

We decided to only show these warnings in development. In production this warning will be handled by Rollbar (because the app as a whole just won't function wihout the authorisation service)