(fix): Add a new Redis instance to address connection errors
- We've been struggling with a bug recently where Redis is seemingly randomly raising connection errors only on Staging
- After a lengthy investigation we've discovered that our configuration
is not following the current advice from Sidekiq to use different
instances Redis for different tasks of caching and queuing [2]. Before
this change we are using a single rdis instance for both caching
geocoding requests and asynchronous jobs.
- We believe that our problem could be explained by a problem with the
instance including latency issues and we want to remove having
everything on the same instance from the investigation.
- Splitting it out also allows us the AWS console to distinguish how many jobs are
in the redis instance compared to cache records, as at the moment it's
just an ever increasing graph as the cache grows
- This change will remove the existing redis instance in order to add
explicit names of 'redis_cache' and 'redis_queue' in the hope that they
will not be confused in the future. The impact of this is that we lose
the outcome of any failing jobs at the time of change, and our
geolocation cache has to be started again. This should be of minimal
impact.
- This change is an important consideration before we ship the work of
job alert nofitications which will be a critical feature using our
queue infrastructure. In the future this change will be much harder as
the impact will be considerably higher
- Both redis instance types can be independently configured/changed so
we can react to problems more easily in the future and scale the job
queue if we need to without impacting the cache redis. This will also
allow us to scale as part of our overall investigation of redis
connection issues if it is a network latency or AWS issue

[1] https://rollbar.com/dxw/teacher-vacancies/items/276/
[2] https://github.com/mperham/sidekiq/wiki/Using-Redis#multiple-redis-instances