(chore) Re-add REDIS_URL for seamless deployment of new Redis clusters
We potentially have a race condition where our app code could be
deployed after Terraform has finished making changes. This would mean
that any references to the old `REDIS_URL` will error for the period of
time where the new code is being deployed.

We should revert this commit in the future once we're happy the new
cache and queue Redis instances are working as expected.