(feature) JSON API results are paginated
We're seeing performance hits to our API because we're returning all
vacancies in one go, this can result in increased server load, long
requests and timeouts.

To prevent the the timeouts and help decrease server load we are now
paginating the set of results returned by the API.

The convention of putting the pagination links in a top-level 'link'
object comes from the JSON API specs.

Added 'self' to the links object as the JSON API says this is also
required.

Changed the query to look for published vacancies instead of not pending
or trashed as it seems more robust (what if we add a new status that
isn't meant to be seen) and takes up less line length :)

Source: https://jsonapi.org/format/#fetching-pagination