(chore) Increase 'per page' number for JSON API
The default 'per page' value for vacancies is set to 10, this is fine
for the front end but for API results we can get away with showing a lot
more. We should increase this to a sensible number that can cope
performance wise. 50 feels like a good place to start, we can always
increase it later.