(chore) Auto scaling supports higher volumes of requests
* Before this change, 10 or more concurrent requests to the root path
would provide enough load to reduce the request time to 7 seconds. With
calls to the API introduced at the same time, some requests start to 503.
* This change suggests the increase from a base of 3 desired web containers to 6. With each container having 256 CPU and 512 memory it has been supporting 5 puma threads. Each instance of the app will also have only 5 default connections to Postgres and Redis. Changing to 6 will change us from 15 threads to a max of 30 available threads.
* Autoscaling has been adjusted to fire up more containers at a lower
threshold to respond to higher CPU use. The maximum containers has
changed from 10 to 20, which should mathmatically be okay as that's 100
Puma threads with 100 connections to postgres which our t2.medium instance can
support with ease (296 max -
https://serverfault.com/questions/862387/aws-rds-connection-limits). This doesn't include the maximum of 6 sidekiq
processes running, each with their own 5 connections to postgres, which
takes us to a maximum of 130.
* After this change the same amount of requests continue to take 1-3 seconds for the home page (which does need caching too).
* set logspout to daemon as it wasn’t always going onto autoscaled instances, sitting at 2/6 which won’t give us logs.
* deployments cannot introduce more than 20 containers into the pool, This ensures there will always be enough EC2 resource to do a deployment while at max load