(chore) Web service is autoscales on lower levels of resource usage
* When high load is detected, scale up sooner. Rather than waiting for 5
full minutes, only wait for 2. This reaction time gets 4 new web
containers in place sooner, reducing chances of 503s and improves the average request time during
short bursts of high requests. There is a balance to be made
between being too sensitive and not reacting in time. I think the
service is being used enough now to justify over provisioning if in
doubt
* CPU usage is generally very low on the web containers, during my tests and review of what the resting CPU might be I think it’s much more likely to be less than 30% for a resting value. Since we use CPU usage as a way to infer increasing work for the service, this was never being triggered.
* Memory wasn’t being monitored before. It’s generally much higher on web, currently on production it’s over 80% at rest. If CPU is 10% but memory is 100%, we should still autoscale. In ECS you can see the competing rules step on each other but you can see that the down alarms respect contradictory alarms for going up and this will scale the service up to 20 containers.