(refactor) Decrease memory allocations of VacanciesHelper
* This was making a enumerating and creating new hash every time the
home page was requested. Using a constant and freezing it tells Ruby
that it only needs one memory allocation.
* Before this change I used derailed_benchmarks to establish that this
line was making 58000000 bytes of memory on 1000 requests, the 7th
highest allocation. After this change the same test reduces the memory
allocation down to 3896000 bytes.
* These values have not changed a year after launch and changing from
ruby to a single hash shouldn't present any issues for us in managing
this data