(fix) SpreadsheetWriter#last_row now calls correct way of returning row
Previously we were calling the `[]` method on the worksheet which
returns a specific cell, therefore requires a row and a column (either
two integers or string eg. 'A2')

We need to call the worksheet's `rows` method in order to return an
array of rows. Then our `#last_row` method will work as expected.