(fix) First column in spreadsheet should be current time job ran at
By having this value as `created_at`, the next time we run the job we
get all the audit data since AND including the last one, so we end up
with the last one from the previous run duplicated.

When we change the value to be the current time, the next time we run
the job we'll get all the audit data since the last time it was ran. So
we no longer duplicate the last audit data record.

This doesn't completely make sense as we're retrieving the records with
a > comparison on the date: `results.where('created_at > ?',
last_updated)` - so you would expect to only get results since the last
audit data was created. In reality we're seeing the last record
included in the result set.

I can't think of a good reason not to use the current time for this
first column at the moment.