(fix) Use the current time for the first column on the spreadsheet
This is a similar change to the one we made for audit data. If we use
`created_at` as the 'last run' comparison we end up getting duplicate
rows, instead we can use the time that the job ran at.

I decided to add the `created_at` data as the last column as I thought
this might be useful data to audit.