(chore) Log sign-in call back event with stored state and Accept header
We're seeing a few occurrences of sign-in callback errors. To help us diagnose
this we need to log the stored state as well as the state that comes back from
DSI.

From looking at the logs we can see multiple requests hitting our callback
endpoint at the same time, the subsequent ones seem to have */* as the value
of the Accept header. This could be caused by erroneous software on the user's
computer, it would be nice to see what values it can be when we see failures.