(chore) Temporarily turn off audit search data Google Sheet job
It's the last remaining Google Sheet job that uses index instead of GID.
We should decide what to do with it then take our time to refactor
out using the index in Google Sheets.

For now it's having unintended consequences overwriting whichever sheet
 is in position 3 and the worksheet's positions seem to randomly
change so we decided to turn it off for now.