(feature) Add JobPosting class to map data back to a Vacancy
To seed the vacancy we need to make a best effort to make sure the
vacancies are valid. For instance, if the publish_on or expires_on dates
are in the past then we need to set them to be today/in the future
respectively.

If we can't find the school by URN for whatever reason then we need to
assign a random school. I used `offset` here instead of `all.sample` as
I read it had better performance, especially when there are alot of rows
in the database.