(feature) Order the API results by publish on date
At the moment API results are returned with no explicit order, meaning
we get back old vacancies in the first page. Some of these vacancies are
now consider invalid as the validation rules have changed over time.

By changing the order to return the most recently published vacancies
first we can ensure the validity and improve the chance that the vacancy
will be saved.