(refactor) Rescue the response errors in the rake task
Instead of rescuing the error in the API wrapper we should allow it to
bubble up to the rake task where we can rescue it and output the errors
to the user executing the task. Otherwise they would probably see a
weird error like NoMethodError on nil or something similar.