(chore) Disable Rails/HasAndBelongsToMany cop for WorkingPattern
Without this cop we would have to create another ActiveRecord model for
the join, then add two relationships to both Vacancy and WorkingPattern
models.

The advantage of using has_many :through is that you can have additional
attributes on the pivot table. However this doesn't apply to us in this
case so it's not worth the extra effort at the moment.

In the future we may want to add additional data to the pivot table and
could remove this exclusion then.