(chore) Log the newly created vacancy's ID on successful import
So that we have a trail of the successful imports, this will help us if
we get in a state where we want to remove imported vacancies.

I had to use a null object double in the test as there were other calls
to `Rails.logger` that were taking precedent over the one I was
expecting.