(feature) Only allow importing vacancies to run on whitelisted DB
We already have some guards against running this rake task in
production. However there is a low chance that the rake task could run
in a production environment and having another guard in place will
give us more confidence in it.

The databases that are on ECS are prefixed with `tvs2_` instead of
`tvs_` as that is what is defined in the Terraform variable
`project_name`. For local development it should always be
`tvs_development`.