(fix) Checkout the git branch before pulling it
`git pull origin $GIT_BRANCH` pulls and merges `$GIT_BRANCH` into
the currently checkout out branch. Doing that before checking out
the intended branch both merges the branch into whatever the
current head is, and leaves the local branch unchanged, meaning we
then deploy a possibly out of date version.

This change ensures we're on the branch we're intending the deploy
before we pull it, updating it to the latest version.