(chore) Implement Vacancy#part_time? now that we don't get it for free
We used to be able to check if a vacancy was part time or not by using
the implicit helpers that enums give us.

We're using `part_time?` to add 'pro rata' to the salary text so it
probably makes sense to only add it if it is equal to 'part_time' and
not 'full_time' as well.

Ideally we'd have a separate field to say if the salary was pro rata or
not.