(refactor) Vacancy attributes now return working patterns as strings
Previously `vacancy.attributes` only returned an array of integers which
throws an error if you do `Vacancy.new(vacancy.attributes)` therefore
we need to merge in the working pattern strings manually. We don't want
to override the `attributes` method as we may need it somewhere else so
we call it `all_attributes`.