(feature) Add search filters for subject and job title
Job title is a simple match filter on the job_title field. It
searches for all given words in the fields with automatic
fuzziness according to Elasticsearch's defaults.

Subject is a more complex query. It searches the three subject
fields in addition to the job title. Querying job title is needed
because subjects are optional, and job title might be the only
place the subject is specified. The query scoring is weighted
towards main subject, then the two supporting subjects, then job
title. It is an "or" query, meaning results containing any of the
search terms will match.

This also refactors the keyword query so we can reuse the fallback
to the null query if the filter is empty.

Elasticsearch documentation on fuzziness:
https://www.elastic.co/guide/en/elasticsearch/reference/current/common-options.html#fuzziness