(feature) Remove keyword filter from search
This refactors keyword filter to only be accessible by alerts by
creating a subclass of `VacancyFilters` called
`VacancyAlertFilters` that extends functionality. The hope is that
one day, we can remove the keyword filter from this, once there
are no existing suibscriptions using it.

This also changes most of the tests to use subject as the primary
filter over keyword, given that is it now the most complex filter.

There is also a slight change in how we test stopwords in
searches. A search for "Part" was being used while a subject
included "Art". With the keyword search, a prefix length is set,
meaning searches for "Art" had to start with "A". The new search
fields are more flexible than that (hoping to catch more typos),
which was causing "Part" searches to find "Art". This is because
the stopwords analyzer (correctly) isn't being used on the subject
fields, so the test was succeeding by coincidence.