(fix) Pin to latest instead of a specific task revisions
My hunch is that our deployment process is creating new task
definitions, but Terraform doesn't pick up on them because it's a
computed property of a task definiton. Because we were pinning to
specific task revisions, we were rolling back to the last revision
Terraform knew about, which was the last revision it created. This
change makes it so that we pin to latest, which will always be the
most recent definition.

The revisions are pinned to latest by default, but are
overridable.