(fix) Use live data for ECS task definitions for service pinning
Terraform `data` resources fetch the current state of the live
system, rather than relying on the Terraform state. This means we
can create new task definitions outside of Terraform and have it
pick them up in a plan. Our build and deploy process does this,
which was causing an issue where `terraform apply`s were rolling
back to earlier versions if there was no change to the task
definition.

This may cause issues on a fresh system deployment, as there is no
matching task definition at that point.

https://github.com/hashicorp/terraform/issues/16380#issuecomment-355846367