(fix) default to nil when fetching OS API key
We were getting some strange errors after deploying to staging because the worker didn't have the API key set. Because it defaults to an empty string, Geocoder treating it like a valid API key, so not raising the expected error, meaning we were getting through to the API and not getting the expected response.