(feature) Create a checkbox collection wrapper
We need to add the item label styles via CSS as `item_label_class` isn't
yet supported by simple_forms for inline labels. Support was added by
plataformatec/simple_form#1637 but is not yet in a released version.
With that change, we could replace the stylesheet changes with
`item_label_class: 'govuk-label govuk-checkboxes__label'` on the
`checkboxes` wrapper.