(feature) Replace enum working_pattern with array_enum working_patterns
This replaces all instances of `working_pattern` on vacancies with
`working_patterns`. Filters still use single select for working pattern,
so are unchanged.

This changes the job specification to have `employementType` as a comma
separated string. https://schema.org/JobPosting specifies the type as
`TEXT`, with no other requirements, so this looks to be an ok change.