(feature) Request all new salaries as full time equivalent
This simplifies the messaging for users, as they always enter the FTE
salary, regardless of working patterns.

We still support displaying and editing existing pro rata salaries as
they were entered, because the guidance was to submit pro rated
salaries if the job was part time.

New vacancies are created without `pro_rata_salary` set.