(feature) Make `weekly_hours` edit only
We show it if it exists and is relevant to the vacancy (i.e. if it's
some sort of flexible working). We can't remove it entirely, because
that would prevent users from being able to edit it. Also, if a user
ever does remove it, they will be unable to add it back in again.