(fix) Require explicit changes to explicitly set flexible working
Vacancies with explicit values set for flexible working will remain as
they are until they match with the derived value, at which point they
will be cleared and will use the derived value henceforth.

Vacancies with explicitly set values for flexible working will display
the checkbox to edit that value when editing the vacancy.