(fix) Change geocoder initializer to use same Redis connection as rest of app
Previously, testing the Geocoders caching behaviour would have required a Redis instance
to be listening on the URL provided.  We would prefer to use MockRedis, as it is a
lightweight alternative to booting up a Redis server during the tests.

Using the main Redis connection requires us to reorder the initilizers such that
the redis_objects ran before geocoder.

This has the side effect of using Redis database 1 for geocoder going forward,
as opposed to 0.

This shouldn't cause any key conflicts with our application as geocoder already
prefixes its cache keys.