(fix) remove data-module attribute from tabs
The data-module attribute on the tab element for vacancy listings on a
school page was encountering a null attribute error from the gov uk
frontend JS.

This fix removes the attribute as the tabs are only used for their
visual appearence. The frontent toolkit JS functionality adds the
ability to switch which panel is visible, but this isn't used as we
reload the panel using AJAX instead.