(fix) content changes and add more JS tests
Modified the yaml file to add the new content changes. These changes
were also tested by new tests added to the
hiring_staff_can_add_stats_to_a_vacancy_spec file to ensure that the
correct content displayed for the given scenarios, as well as enabling
javascript for the tests as it was not enabled previously.