(fix) Skip validation when saving vacancy feedback
Vacancies which were previously valid, but are no longer because
of newer validations being added, mean that using the update method
on this model can return false.

We don't care that the vacancy is no longer valid according to the
current definition of "valid". Only that we're recording what the
feedback on a particular vacancy was.