Fix typo in the category declaration
The category in the job was declared as 'vacancy' and when we ran this particular job, the job would return an error that a method called num_rows could not be called on NIL. 'num_rows' works on a declared worksheet variable so this means that the worksheet itself was nil.

Upon investigating why the worksheet could not be found, (since the spreadsheet IDs and GIDs matched up with the relevant spreadsheets) it was discovered that the category associated with the GID of the sheet we were concerned with was defined as 'vacancies' in the audit data model.

The category being passed to the job was the singular 'vacancy' which did not exist nor have a GID associted with it and therefore no worksheet could be fetched.

When the GID was NIL, this was falling back to zero in development and working because there was a development sheet with GID zero, but in production there is no sheet with GID zero.

This change fixes this.