(fix) change org link display to users
The change organisation link will now only display to user with access
to multiple schools, regardless of whether the dfe sign-in feature flag
is toggled on or off. We had to change the test to reflect this new
behaviour as the sessions controller was modified to detect if a user
has access to multiple schools regardless of the authorisation method
used for users.