(chore) Add user reference and foreign key to vacancies
We added this because we want to be able to match the vacancy to the user who created it. So that we can send email prompts to request feedback on expired vacancies older than 2 weeks.

We also changed the name of the newly added user_id column to publisher_user_id as a way to describe the particular user we want match the published vacancy to. In this case, this would be the user who presses the submit button on the review job details page (i.e performs the final call to action to publish the vacancy).