Save id of user who performed submit action
We want to save this id as the publisher_user_id for a publsihed vacancy in the vacancy model.
The current user method is defined in the BaseController from which the publish controller can inherit and pass as variable to the Publish Vacancy class.

It is from this class that we can obtain and set this user information at the point of vacancy being published.