Edit general feedback page
This HAML file initially shared a partial that uses the 'now removed' rating attribute.
Since another HAML file still uses this partial, the partial itself was not edited.
Since this redesign is only for this feedback page, the partial was removed instead and it's content copied over and edited to suit the new design of this page.

A new styling was added as the previous styling (which is also used elsewhere) resulted in large gaps between the header and the textbox after removing the rating options.