Indent the textbox to correct styling
Custom CSS was initally used to reduce the gap between the textbox and its header.
It was later pointed out from review that the large spacing was due to the textbox being indented incorrectly in this HAML file.
This change indents the line to fall within the fieldset component and removes the custom css added previously.