Remove unnecessary tests for content
Testing static content could give confidence that the test setup has successfully navigated to the page of choice. e.g if the route was to change, then the tests would fail.
However, this can be done by just checking that the page heading is for the page we want.
The extra content tests are unnecessary and in this change, removed.