Remove rating from general feedback factory
This is a quick change to remove the rating attribute from the general feedback factory.
It is no longer needed.
It also shows that we get nil in the rating column (a.k.a an empty cell) when we write to the audit spreadsheet. Showing that these changes to remove the rating should not have an affect on successfully writing to the audit spreadsheet.