Add new columns to general feedback model.
The second slice of the general feedback page redesign involves adding a section to ask users whether they would like to participate in further user research sessions.The user would select either yes or no. If yes, then we would request an email to contact.

To prepare for this, an email column and a response column were added to the general feedback model.