(chore) Add canonical link to head of page
As part of SEO work the vacancy listings pages need to have canonical
links in the head so that they can be indexed properly by search
engines. This uses a partial to render content into the header of the
vacancy listings pages, along with prev/next rel link tags as these were
suggested to help with structuring related to paginated content.