Make email required when user is interested.
In this change, we added validation for the presence of an email when the user is interested in participating in further user research, i.e. selects 'Yes' as their response.

It also adds an email field underneath the 'Yes' radio button on the form and passes email as part of the general feedback params.