Unroll collection radio buttons
In order to add an email text field below the 'Yes' radio button only, we separated the collection of radio buttons into individual radio buttons.

If this was not done, the email text field would have appeared under both radio buttons.

We also removed the helper method for the participation options as this was no longer needed.