Refactor enums in codebase into hash enums.
This is a change is a code clean up across the codebase.
Based on a comment on a PR: "It is best practice to add a specific integer value for each enum value."

Not all the enums declared in the codebase are following this best practice and this could lead to more enums being created (especially in a copy & paste scenario) that don;t follow this practice.

This change was done to encourage good code practices as mentioned above.