(fix) add GOV.UK to title as suggested
The accessibility audit performed highlighted that page titles weren't
descriptive enough for screen readers - the feedback given was “I found
that the page title did not include company information after describing
the purpose of the page. This information may be important for some
users when navigating multiple tabs.” so this adds GOV.UK to the title
of every page of the service using localization.