Update back button to support redirect GO experiments
GO redirect tests are redirecting the user to a new URL. The
back link would take them to the page before the redirection.
Changing the back link to use JS method will ensure that
the user is being navigated to the right page. Updated the tests
to ensure this logic supports both JS enabled/disabled.