(fix) change gem to fixed version
The gov-uk-date-fields gem is used to generate the date fields that the
hiring staff fill out when creating a vacancy. An accessibility audit
found that the label for the month input fields was not actually
associated with the date fields. This was because the for attribute was
being written in to the class by mistake (by the gem), so we fixed the
gem in a fork then changed the gemfile to refer to that version instead
as a temporary solution.