Update navigation header item
(chore) add links and styling to footer
As part of work to improve the site on mobile this adds some additional
links to the footer as well as some additional styling

(chore) move navigation to header
This change moves the navigation to the header and make the navigation
responsive using the GOVUK frontend toolkit. It makes a menu button
appear on mobile devices to toggle the dispay of the navbar.

(chore) add extra footer links and header labels
As part of improving navigation, navigation links required in the footer
Updated the i18n yml accordingly