Use  original gov_uk_date_fields gem
Prior to this change, we were using a forked version of the above gem.
The forked version introduced gov uk styling which was absent in the original gem at the time it was used.
Since then, there has been a version release of the original gem that uses styling from the GDS design system.

From a maintenance standpoint, we want to use the original gem so that we don't miss any fixes or updates.