(fix) upgrade frontend toolkit to 3.0.0
This commit upgrades the frontend toolkit to version 3.0.0 as we want
to have the accessibility changes that were added to the toolkit. This
commit also fixes any breaking changes that caused the tests in the test
suite to fail. The styling on some pages needs to be amended to use the
changes the latest toolkit version introduced.