(chore) Change vacancy model to handle expiry time
Changed the scope quieries in the vacancy model, so that
expiry_time is used in quieries when its not nil, while
tolerates the absence of this value. So that existing vacancies
that has no expiry_time can be expired based on their expies_on date,
just like the old behaviour. Added tests accordingly.
Disable Rubocop to check TimeZone so that we can stub expiry time
without having a defined timezone