Replace sort links with sort select options
The job listings on the search page can be sorted based on their closing dates and published dates.
The previous design involved using sorting links that redirected the user to a page with the sorted content.

As part of a design change to improve the UX. These links have been replaced with more descriptive select options.
This change also involves the use of both a rails form and custom javascript for better accessibilty.

For javascript enabled browsers, the user just needs to select their preferred filter option and the page would redirect to the sorted content.

For javascript disabled browsers, the user would need to select their option and then click a button (Only visible if javascript is disabled) to submit a 'form' which responds by redirecting them to their sorted content.