Add spec for sorting in JS enabled browsers
The majority of the tests in this feature spec were written from the perspective of a user usinf a JS DISABLED browser.

There is javascript that hides the sort button from the user's view and submits the form automatically when the user selects an option.

This change adds a spec to test the feature from the perspective of a JS enabled browser.