Add clear all data method to spreadsheet writer
We want to use the spreadsheet writer to write the details of DSI users within Teaching vacancies' scope to a specific spreadsheet.

In order to keep the data in the spreadsheet in sync with the data in the DSI database, a decision was made to clear the spreadsheet first and then add the data pulled from the DSI API on a daily basis.

This commit adds a method that would allow us to clear all the data in a spreadsheet.