WIP Add class to request user data per page
The class would make a request to the API (through the DFE:SignIn:API class) for the desired user data.
It would make as many requests as there are pages. The total page number is obtained from an initial request to the API.

It would then add this data to a specified spreadsheet and implement some methods defined in the spreadsheet writer class.

This is still WIP as it is may be desired that this class would be able to do the following:
- Initialize with a particular spreadsheet
- Use the spreadsheet writer clear_all_rows method to clear the entire worksheet before appending new data
- Handle any errors with bad requests/responses for specific pages

The above have not yet been implemented and may be done in follow up commits or in this one.