Add use current location functionality
In this work we use Postcodes.io to retrieve the postcode from the 
coordinates we receive from the browser. This way we avoid to implement
an endpoint on the server and reduce the load we have to handle.
Testing of the functionality is done using mockjax library because we
cannot simply stub a request from the server with webmock.