(chore) remove unneeded filter classes and styling
After consulting with the UX designer it was agreed that the grey border
didn't look good with the new focus state of the govuk frontend. The
form-control class on the search filters no longer serves a purpose
since the css for it was removed - nothing else refers to this class so
it is safe to remove it. The form-control-block wasn't used by anything
either so it was also removed to clean up the code.